+++ Erg_1/app/src/main/java/com/example/expense_recording/Interfaces/UserManager.java
@@ -8,13 +8,13 @@ public interface UserManager {
     // download data from store
     void set_date_from_store();
     void change_date(int day, int month, int year);
-    void add_record(int record);
+    void add_record(float record);
 
     // methods for statistics
-    int get_average(Date d);
-    int get_max(Date d);
-    int get_min(Date d);
-    int total(Date d);
+    float get_average(Date d);
+    float get_max(Date d);
+    float get_min(Date d);
+    float total(Date d);
 
     int getDay();
     int getMonth();
